Revert "Revert "correction bug""
This reverts commit 3a60822d573007b8fd621bc0a5e68113352cd0d6.

+++ db/MasObject.py
@@ -440,8 +440,8 @@ class calcul_abscisse(MasObject):
             RETURNS trigger AS  
             $BODY$ 
             DECLARE  
-                long1	float; 
-                long2	float; 
+                long1	double precision; 
+                long2	double precision;  
                 g	geometry; 
                 b	integer; 
                 z	integer; 
@@ -471,7 +471,7 @@ class calcul_abscisse(MasObject):
                            long1 = 0;
                        END IF;
                         
-                       NEW.abscissa = ROUND(long1+long2,1);
+                       NEW.abscissa = ROUND((long1+long2)::numeric,1);
 
                     END IF;
 
@@ -494,8 +494,8 @@ class calcul_abscisse(MasObject):
                   RETURNS trigger AS
                 $BODY$
                 DECLARE
-                    long1	float; 
-                    long2	float; 
+                    long1	double precision;
+                    long2	double precision;
                     g	geometry;
                     p	geometry;
                     b	integer;
@@ -519,7 +519,7 @@ class calcul_abscisse(MasObject):
                             long1 = 0;
                         END IF;
                         
-                        NEW.abscissa = ROUND(long1+long2,1);
+                        NEW.abscissa = ROUND((long1+long2)::numeric,1);
 
                     END IF;
                         
@@ -548,8 +548,8 @@ class calcul_abscisse(MasObject):
                     long1 = 0; 
                 END IF; 
             
-                NEW.zoneabsstart = ROUND(long1,1);
-                NEW.zoneabsend = ROUND(long1+long2,1);
+                NEW.zoneabsstart = ROUND(long1::numeric,1);
+                NEW.zoneabsend = ROUND((long1+long2)::numeric,1);
             
                 RETURN NEW;
             END;
